+++ sim/game/cards/dm09/beast_folk.go
@@ -61,13 +61,13 @@ func StormWranglerTheFurious(c *match.Card) {
 		fx.When(fx.InTheBattlezone, func(card *match.Card, ctx *match.Context) {
 			ctx.Match.ApplyPersistentEffect(func(ctx2 *match.Context, exit func()) {
 				if card.Zone != match.BATTLEZONE {
-					card.RemoveConditionBySource(card.ID)
+					card.RemoveSpecificConditionBySource(cnd.PowerAmplifier, card.ID)
 					exit()
 					return
 				}
 
 				if _, ok := ctx2.Event.(*match.EndOfTurnStep); ok {
-					card.RemoveConditionBySource(card.ID)
+					card.RemoveSpecificConditionBySource(cnd.PowerAmplifier, card.ID)
 					exit()
 					return
 				}